// transporte-web/src/app/core/models/vehicle.ts
export interface Vehicle {
    id: number;
    plate: string;
    engine: string;
    chassis: string;
    model: string;
    registrationDate: string;
    seatedPassengers: number; 
    standingPassengers: number;
    dryWeight: number;
    grossWeight: number;
    doors: number;
    brand: string;
    line: string;
}